+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed being able to drain energy from the Refined Storage Controller (raoulvdberge)
 - Fixed the Grid crashing on a item/fluid update-heavy storage system (raoulvdberge, Darkere, noobanidus)
 - Fixed crash with the Disk Manipulator and using item/fluid filters when inserting into the network (raoulvdberge)
+- Fixed the network being able to run off 1 FE/t (raoulvdberge)
 
 ### 1.7.1
 - Exact mode for processing patterns no longer exist, you can now define per-slot which item/fluid tags are allowed to be used by autocrafting, by using CTRL + click on the filter slot in the Pattern Grid (raoulvdberge)